requirements.txt
Django~=2.2.4
django-filters~=0.2.1
django-taggit~=1.2.0